+++ src/cell/visor.c
@@ -68,6 +68,7 @@ visor_begin_op(visor_t* visor, const char* fmt, ...)
 	print_indent(visor->indent_level);
 	vprintf(fmt, ap);
 	printf("...\n");
+	fflush(stdout);
 	va_end(ap);
 
 	++visor->indent_level;
@@ -89,6 +90,7 @@ visor_error(visor_t* visor, const char* fmt, ...)
 	printf("E: ");
 	vprintf(fmt, ap);
 	printf("\n");
+	fflush(stdout);
 	va_end(ap);
 
 	++visor->num_errors;
@@ -103,6 +105,7 @@ visor_print(visor_t* visor, const char* fmt, ...)
 	print_indent(visor->indent_level);
 	vprintf(fmt, ap);
 	printf("\n");
+	fflush(stdout);
 	va_end(ap);
 }
 
@@ -116,6 +119,7 @@ visor_warn(visor_t* visor, const char* fmt, ...)
 	printf("W: ");
 	vprintf(fmt, ap);
 	printf("\n");
+	fflush(stdout);
 	va_end(ap);
 
 	++visor->num_warns;