+++ src/plugin/NativeMethods.cs
@@ -9,8 +9,7 @@ namespace miniSphere.Gdk
 {
     static class NativeMethods
     {
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
+        [DllImport("user32.dll")] [return: MarshalAs(UnmanagedType.Bool)]
         internal static extern bool SetForegroundWindow(IntPtr hWnd);
     }
 }